refactor: ProductSearchService를 ProductScheduleQueryService에 통합하고 컨트롤러 의존성 및 호출 로직 정리
