+++ src/main/java/com/yeoljeong/tripmate/product/presentation/controller/internal/ProductFeignClientController.java
@@ -1,7 +1,7 @@
 package com.yeoljeong.tripmate.product.presentation.controller.internal;
 
 import com.yeoljeong.tripmate.product.application.service.query.ProductQueryService;
-import com.yeoljeong.tripmate.product.application.service.query.ProductSearchService;
+import com.yeoljeong.tripmate.product.application.service.query.ProductScheduleQueryService;
 import com.yeoljeong.tripmate.product.presentation.dto.response.ProductScheduleInfoResponse;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
@@ -15,27 +15,26 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/internal/products")
 public class ProductFeignClientController {
 
-  private final ProductQueryService productQueryService;
-  private final ProductSearchService productSearchService;
+  private final ProductScheduleQueryService productScheduleQueryService;
 
-  //일정 확정시 사용 하는 상품 정보
+  // 일정 확정시 사용 하는 상품 정보
   @GetMapping("/schedules/{scheduleId}")
   public ProductScheduleInfoResponse getSchedule(
       @PathVariable UUID scheduleId
   ) {
     return ProductScheduleInfoResponse.from(
-        productSearchService.getSchedule(scheduleId)
+        productScheduleQueryService.getSchedule(scheduleId)
     );
   }
 
-  //주문시 사용 하는 상품,스케줄 정보
+  // 주문시 사용 하는 상품, 스케줄 정보
   @GetMapping("/{productId}/schedules/{scheduleId}")
-  public ProductScheduleInfoResponse getSchedule(
+  public ProductScheduleInfoResponse getProductSchedule(
       @PathVariable UUID productId,
       @PathVariable UUID scheduleId
   ) {
     return ProductScheduleInfoResponse.from(
-        productSearchService.getProductSchedule(productId, scheduleId)
+        productScheduleQueryService.getProductSchedule(productId, scheduleId)
     );
   }
 }
